+++ src/QueueCommandString.php
@@ -5,6 +5,41 @@ namespace Laravel\Horizon;
 class QueueCommandString
 {
     /**
+     * Get the additional option string for the worker command.
+     *
+     * @param  \Laravel\Horizon\SupervisorOptions  $options
+     * @return string
+     */
+    public static function toWorkerOptionsString(SupervisorOptions $options)
+    {
+        return sprintf('--name=%s --supervisor=%s %s',
+            $options->workersName,
+            $options->name,
+            static::toOptionsString($options)
+        );
+    }
+
+    /**
+     * Get the additional option string for the supervisor command.
+     *
+     * @param  \Laravel\Horizon\SupervisorOptions  $options
+     * @return string
+     */
+    public static function toSupervisorOptionsString(SupervisorOptions $options)
+    {
+        return sprintf('--workers-name=%s --balance=%s --max-processes=%s --min-processes=%s --nice=%s --balance-cooldown=%s --balance-max-shift=%s %s',
+            $options->workersName,
+            $options->balance,
+            $options->maxProcesses,
+            $options->minProcesses,
+            $options->nice,
+            $options->balanceCooldown,
+            $options->balanceMaxShift,
+            static::toOptionsString($options)
+        );
+    }
+
+    /**
      * Get the additional option string for the command.
      *
      * @param  \Laravel\Horizon\SupervisorOptions  $options
@@ -13,9 +48,9 @@ class QueueCommandString
      */
     public static function toOptionsString(SupervisorOptions $options, $paused = false)
     {
-        $string = sprintf('--delay=%s --memory=%s --queue="%s" --sleep=%s --timeout=%s --tries=%s',
-            $options->delay, $options->memory, $options->queue,
-            $options->sleep, $options->timeout, $options->maxTries
+        $string = sprintf('--backoff=%s --max-time=%s --max-jobs=%s --memory=%s --queue="%s" --sleep=%s --timeout=%s --tries=%s',
+            $options->backoff, $options->maxTime, $options->maxJobs, $options->memory,
+            $options->queue, $options->sleep, $options->timeout, $options->maxTries
         );
 
         if ($options->force) {