+++ src/RedisQueue.php
@@ -56,11 +56,30 @@ class RedisQueue extends BaseQueue
      */
     public function pushRaw($payload, $queue = null, array $options = [])
     {
-        $payload = (new JobPayload($payload))->prepare($this->lastPushed)->value;
+        $payload = (new JobPayload($payload))->prepare($this->lastPushed);
 
-        return tap(parent::pushRaw($payload, $queue, $options), function () use ($queue, $payload) {
-            $this->event($this->getQueue($queue), new JobPushed($payload));
-        });
+        parent::pushRaw($payload->value, $queue, $options);
+
+        $this->event($this->getQueue($queue), new JobPushed($payload->value));
+
+        return $payload->id();
+    }
+
+    /**
+     * Create a payload string from the given job and data.
+     *
+     * @param  string  $job
+     * @param  string  $queue
+     * @param  mixed  $data
+     * @return array
+     */
+    protected function createPayloadArray($job, $queue, $data = '')
+    {
+        $payload = parent::createPayloadArray($job, $queue, $data);
+
+        $payload['id'] = $payload['uuid'];
+
+        return $payload;
     }
 
     /**
@@ -155,15 +174,5 @@ class RedisQueue extends BaseQueue
                 $event->connection($this->getConnectionName())->queue($queue)
             );
         }
-    }
-
-    /**
-     * Get a random ID string.
-     *
-     * @return string
-     */
-    protected function getRandomId()
-    {
-        return JobId::generate();
     }
 }