+++ src/SupervisorOptions.php
@@ -2,9 +2,7 @@
 
 namespace Laravel\Horizon;
 
-use Illuminate\Queue\WorkerOptions;
-
-class SupervisorOptions extends WorkerOptions
+class SupervisorOptions
 {
     /**
      * The name of the supervisor.
@@ -12,6 +10,13 @@ class SupervisorOptions extends WorkerOptions
      * @var string
      */
     public $name;
+
+    /**
+     * The name of the workers.
+     *
+     * @var string
+     */
+    public $workersName;
 
     /**
      * The queue connection that should be utilized.
@@ -77,13 +82,73 @@ class SupervisorOptions extends WorkerOptions
     public $balanceMaxShift = 1;
 
     /**
+     * The number of seconds to wait before retrying a job that encountered an uncaught exception.
+     *
+     * @var int
+     */
+    public $backoff;
+
+    /**
+     * The maximum number of jobs to run.
+     *
+     * @var int
+     */
+    public $maxJobs;
+
+    /**
+     * The maximum number of seconds a worker may live.
+     *
+     * @var int
+     */
+    public $maxTime;
+
+    /**
+     * The maximum amount of RAM the worker may consume.
+     *
+     * @var int
+     */
+    public $memory;
+
+    /**
+     * The maximum number of seconds a child worker may run.
+     *
+     * @var int
+     */
+    public $timeout;
+
+    /**
+     * The number of seconds to wait in between polling the queue.
+     *
+     * @var int
+     */
+    public $sleep;
+
+    /**
+     * The maximum amount of times a job may be attempted.
+     *
+     * @var int
+     */
+    public $maxTries;
+
+    /**
+     * Indicates if the worker should run in maintenance mode.
+     *
+     * @var bool
+     */
+    public $force;
+
+
+    /**
      * Create a new worker options instance.
      *
      * @param  string  $name
      * @param  string  $connection
      * @param  string  $queue
+     * @param  string  $workersName
      * @param  string  $balance
-     * @param  int  $delay
+     * @param  int  $backoff
+     * @param  int  $maxTime
+     * @param  int  $maxJobs
      * @param  int  $maxProcesses
      * @param  int  $minProcesses
      * @param  int  $memory
@@ -95,22 +160,43 @@ class SupervisorOptions extends WorkerOptions
      * @param  int  $balanceCooldown
      * @param  int  $balanceMaxShift
      */
-    public function __construct($name, $connection, $queue = null, $balance = 'off',
-                                $delay = 0, $maxProcesses = 1, $minProcesses = 1, $memory = 128,
-                                $timeout = 60, $sleep = 3, $maxTries = 0, $force = false, $nice = 0,
-                                $balanceCooldown = 3, $balanceMaxShift = 1)
+    public function __construct($name,
+                                $connection,
+                                $queue = null,
+                                $workersName = 'default',
+                                $balance = 'off',
+                                $backoff = 0,
+                                $maxTime = 0,
+                                $maxJobs = 0,
+                                $maxProcesses = 1,
+                                $minProcesses = 1,
+                                $memory = 128,
+                                $timeout = 60,
+                                $sleep = 3,
+                                $maxTries = 0,
+                                $force = false,
+                                $nice = 0,
+                                $balanceCooldown = 3,
+                                $balanceMaxShift = 1)
     {
         $this->name = $name;
-        $this->nice = $nice;
-        $this->balance = $balance;
         $this->connection = $connection;
+        $this->queue = $queue ?: config('queue.connections.'.$connection.'.queue');
+        $this->workersName = $workersName;
+        $this->balance = $balance;
+        $this->backoff = $backoff;
+        $this->maxTime = $maxTime;
+        $this->maxJobs = $maxJobs;
         $this->maxProcesses = $maxProcesses;
         $this->minProcesses = $minProcesses;
-        $this->queue = $queue ?: config('queue.connections.'.$connection.'.queue');
+        $this->memory = $memory;
+        $this->timeout = $timeout;
+        $this->sleep = $sleep;
+        $this->maxTries = $maxTries;
+        $this->force = $force;
+        $this->nice = $nice;
         $this->balanceCooldown = $balanceCooldown;
         $this->balanceMaxShift = $balanceMaxShift;
-
-        parent::__construct($delay, $memory, $timeout, $sleep, $maxTries, $force);
     }
 
     /**
@@ -187,14 +273,17 @@ class SupervisorOptions extends WorkerOptions
             'balance' => $this->balance,
             'connection' => $this->connection,
             'queue' => $this->queue,
-            'delay' => $this->delay,
+            'backoff' => $this->backoff,
             'force' => $this->force,
             'maxProcesses' => $this->maxProcesses,
             'minProcesses' => $this->minProcesses,
             'maxTries' => $this->maxTries,
+            'maxTime' => $this->maxTime,
+            'maxJobs' => $this->maxJobs,
             'memory' => $this->memory,
             'nice' => $this->nice,
             'name' => $this->name,
+            'workersName' => $this->workersName,
             'sleep' => $this->sleep,
             'timeout' => $this->timeout,
             'balanceCooldown' => $this->balanceCooldown,